refactor code into MoorhenNotification
